moving stats to worker submodule
